server/requirements.txt
flask==3.1.1
flask-cors==5.0.1
python-dotenv==1.1.0
werkzeug==3.1.3
pyjwt==2.8.0
